% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterGoanaOutput.R
\name{filterGoanaOutput}
\alias{filterGoanaOutput}
\title{Filter annotation enrichment output from goana}
\usage{
\code{filterGoanaOutput(goanaResult, idCol = "GO_ID", orderBy = "P.DE")}
}
\arguments{
\item{goanaResult}{a data frame containing the results of call to \code{goana}}

\item{idCol}{character or numeric, the column of \code{goanaResult} containing the ontology identifiers. Defaults to "GO_ID"}

\item{orderBy}{character or numeric, the column of \code{goanaResult} that contains information on the enrichment of the term. This is passed to \code{dplyr::arrange}. Defaults to "P.DE"}
}
\value{
A data frame that is a subset of the input \code{goanaResult}, retaining only rows for terms that are more strongly enriched than any descendant terms
}
\description{
This function filters the output of \code{goana} to include only terms that are either more specific or more strongly enriched. It uses the ancestor-dependent relationships of the Gene Ontology database to identify terms that are less specific, and culls them unless they are more strongly enriched in the analysis than the descendant term.
}
