% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_topGenes_combined.R
\name{get_topGenes_combined}
\alias{get_topGenes_combined}
\title{Construction of combined topGenes object from a set of pairwise contrasts in limma models}
\usage{
\code{
get_topGenes_combined(topGenes,
                      topGenes.pairwise,
                      p_cut=0.01, fc_cut=log2(1.5))}
}
\arguments{
\item{topGenes}{a data frame, typically the output of a call to \code{topTable}. Must contain genes, log2 fold-change, and adjusted p-values.}

\item{topGenes.pairwise}{a list of data frames, each typically containing the output of a call to \code{topTable} for a single contrast. Each list element should be named with an identifier for the contrast, and must contain genes, log2 fold-change, and adjusted p-values. Can optionally include a "threshold" column, which should be boolean indicating genes passing significance thresholds.}

\item{p_cut}{numeric, the cutoff for adjusted p-value. Genes with adjusted p-values greater than or equal to this value are not considered significant.}

\item{fc_cut}{numeric, the absolute value cutoff for log2 fold change. Genes with absolute value log2-FC less than or equal to this value are not considered significant.}
}
\value{
A data frame, with adjusted p-values and logical \code{threshold} values for all contrasts in \code{topGenes.pairwise}
}
\description{
This function takes a set of pairwise limma model contrasts, and combines the results into a single
object.
}

