% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_xy_lims.R
\name{get_xy_lims}
\alias{get_xy_lims}
\title{Accessory function for determining x and y limits for volcano plots}
\usage{
\code{
get_xy_lims(topGenes,
            pairwise=FALSE,
            x_symmetric=TRUE,
            adjust_factor=0.01,
            y_floor_zero=TRUE,
            x_col="logFC", y_col="adj.P.Val",
            neg_log10_ycol=TRUE,
            min_x_abs=NULL,
            min_y2=NULL)}
}
\arguments{
\item{topGenes}{a data frame or list of data frame. Each data frame typically contains the output of a call to \code{topTable}. Must contain genes, log2 fold-change, and adjusted p-values.}

\item{pairwise}{logical, whether \code{topGenes} contains multiple comparisons. Defaults to FALSE. If TRUE, the function finds common limits for all plots based on the most extreme values across all comparisons.}

\item{x_symmetric}{logical, whether the x-axis should be made symmetric.}

\item{adjust_factor}{numeric, the factor for adjusting the limits of the x- and y-axes. The returned limits will be extended by this much to allow points at the extremes of the plot to fit. Defaults to 0.01.}

\item{y_floor_zero}{logical, whether to set the lower limit of y to 0. Defaults to TRUE.}

\item{x_col, y_col}{character or integer, the name or number of the columns containing the data to be plotted on the x- and y-axes. Default to "logFC" and "adj.P.Val", respectively.}

\item{neg_log10_ycol}{logical, whether the y-axis will be plotted as \code{(-log10(y))}. Defaults to TRUE, as the function is intended for use with p-values.}

\item{min_x_abs}{numeric, the minimum absolute value for the extent of the x-axis. Defaults to NULL, which sets no default. Typically used when plotting with thresholds, to ensure that threshold value is shown.}

\item{min_y2}{numeric, the minimum value for upper limit the y-axis. Defaults to NULL, which sets no default. Typically used when plotting with thresholds, to ensure that threshold value is shown.}
}
\value{
A list, with vectors for x- and y-limits.
}
\description{
This function determines appropriate x and y limits for volcano plots or sets of volcano plots.
}
\details{
This function finds reasonable x- and y-limits for plots; it is specifically designed for use with volcano plots.
}
