% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano_2var.R
\name{plot_volcano_2var}
\alias{plot_volcano_2var}
\title{Generate a volcano plot of genes from a differential expression (limma) analysis}
\usage{
\code{
plot_volcano_2var(
     topGenes, my_cols=c("darkcyan", "darkorange"),
     file_prefix=NULL, plotdims=c(9,9),
     color_by_threshold=TRUE, fc_cut=log2(1.5), p_cut=0.01,
     x_lim="auto", y_lim="auto",
     gene_labs=NULL, x_cut=0, y_cut=0, x_cut_direction="both",
     gene_labs_repel=TRUE, gene_lab_size=3,
     ...)}
}
\arguments{
\item{topGenes}{a data frame, typically the output of a call to \code{topTable}. Must contain genes, log2 fold-change, and adjusted p-values. Can optionally include a "threshold" column, which should be boolean indicating genes passing significance thresholds.}

\item{my_cols}{a vector of colors for plotting points. If \code{color_by_threshold} is FALSE, only the first element is used. Otherwise, the first and second elements provide the color for points not exceeding and exceeding significance thresholds, respectively.}

\item{file_prefix}{a character string. If provided, the function outputs a pdf of the plot, named "{file_prefix}.pdf".}

\item{plotdims}{a numeric vector, the size (in inches) of the plotting object. Either the size of the pdf, or the size of the plotting window.}

\item{color_by_threshold}{logical, whether to color points based on exceeding a threshold for logFC and p-value. Used only if fc_cut and p_cut are not NULL.}

\item{fc_cut}{numeric, the (absolute value) log2 fold-change threshold for determining significance of genes. This value is also plotted as vertical dotted lines. Setting to NULL removes the lines.}

\item{p_cut}{numeric, the p-value threshold for determining significance of genes. This value is also plotted as a horizontal dotted line. Setting to NULL removes the lines.}

\item{x_lim, y_lim}{either "auto", NULL, or numeric vectors. If "auto", x- and y-limits are determined from the data using \code{get_xy_lims}. If NULL, default plot limits are used. If provided as numeric vectors, the lower and upper limits of the plotting space along the x- and y-axes. Passed to \code{ggplot2::xlim}.}

\item{gene_labs}{character, the type of gene labeling to include. If \code{"threshold"}. If \code{"ellipse"}, genes with values outside the labeling ellipse will be labeled. Default value is NULL, which yields no gene labeling.}

\item{x_cut, y_cut}{numeric, the radii of the labeling ellipse along the x- and y-axes. Genes with values outside the ellipse are labeled with gene names. Default to 0, which results in all genes being labeled.}

\item{x_cut_direction}{character. For threshold-based gene labeling only, specifies the direction of the threshold for gene labeling. Default to "both", which labels genes with absolute value logFC greater than x_cut. "lower" labels genes with logFC less than x_cut; "upper" labels genes with logFC greater than x_cut.}

\item{gene_labs_repel}{logical, whether to force separation of the gene label text. If TRUE, labels are plotted using \code{geom_text_repel}.}

\item{gene_lab_size}{numeric, the size of the gene label text. Passed to \code{geom_text} to \code{geom_text_repel}.}

\item{...}{additional parameters passed to \code{pdf}.}
}
\description{
Generate a volcano plot of genes from a differential expression (limma) analysis. This plot can be
output to a plotting window, or to a pdf. The points can be colored based on fold-change and p-value
thresholds. Points can also be labeled with gene names, and the points to be labeled can be set based
on an ellipse oriented to the x- and y-axes.
}

