% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_EList.R
\name{trim_EList}
\alias{trim_EList}
\title{Drop samples from EList object}
\usage{
\code{trim_EList(EList, keep_samples, drop_samples=NULL)}
}
\arguments{
\item{EList}{an \code{EList} object containing expression values}

\item{keep_samples}{a numeric vector of the indices of samples to keep; alternately, a logical vector with length equal to number of libraries in EList, indicating which libraries to keep.}

\item{drop_samples}{a numeric vector of the indices of samples to drop; alternately, a logical vector with length equal to number of libraries in EList, indicating which libraries to drop. Ignored if \code{keep_samples} is provided.}
}
\value{
an \code{EList} object with \code{drop_samples} removed from all components, or only \code{keep_samples} retained in all components.
}
\description{
This function drops samples from an EList object, iterating over all components of the object. It uses the dimensions of the components to determine whether to drop rows, columns, etc., and thus will have difficulty if the number of samples is small.
}
\details{
Either \code{keep_samples} or \code{drop_samples} must be specified.
}

