% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_sig_genes.R
\name{write_sig_genes}
\alias{write_sig_genes}
\title{Output text files with lists of significant genes}
\usage{
\code{
write_sig_genes(
  topGenes, file_prefix,
  method=c("ranked_list", "combined", "directional"),
  adj_p_cut=0.01, fc_cut=log2(1.5), fc_adj_factor=1,
  p_col="P.Value", adj_p_col="adj.P.Val", fc_col="logFC",
  threshold_col=NULL)}
}
\arguments{
\item{topGenes}{a data frame, typically the output of a call to \code{topTable}. Must contain genes, log2 fold-change, and adjusted p-values.}

\item{file_prefix}{name of the destination for files. Details of each output will be appended to this prefix.}

\item{method}{character, specifying the type of gene lists to output. "ranked_list" outputs a list of all genes in ranked order by p-value (from smallest to largest). "combined" outputs a list of all significant genes meeting the threshold. "directional" outputs lists of significant genes meeting the threshold that are up- and down-regulated. Partial matches are allowed.}

\item{adj_p_cut}{numeric, the cutoff for adjusted p-value. Genes with adjusted p-values greater than or equal to this value are not included in the result. Defaults to 0.01.}

\item{fc_cut}{numeric, the absolute value cutoff for log2 fold change. Genes with absolute value log2-FC less than or equal to this value are not included in the result. Defaults to log2(1.5). To include all genes, set to 0.}

\item{fc_adj_factor}{numeric, the adjustment factor used for log2-fold-change values with a numeric predictor. This is included so that the output file names can include the fold change prior to scaling. Defaults to 1, which is the appropriate value for categorical comparisons.}

\item{p_col}{name or number of the column in \code{topGenes} on which to sort. Generally the raw p-values, as adjusted p-values are often homogenized across a range of raw p-values. Defaults to "P.Value", which corresponds to the output from \code{topTable}. To include all genes, set to >1.}

\item{adj_p_col}{name or number of the column in \code{topGenes} containing the p-values to compare to \code{p_cut}. Defaults to "adj.P.Val", which corresponds to the output from \code{topTable}.}

\item{fc_col}{name or number of the column in \code{topGenes} containing the fold-change values to compare to \code{fc_cut}. Defaults to "logFC", which corresponds to the output from \code{topTable}.
#' @param threshold_col name or number of the column in \code{topGenes} containing the logical values indicating which genes meet thresholds. This is an alternate way to determine signficance of genes. If specified, \code{p_cut} and \code{fc_cut} are ignored.}
}
\description{
This function outputs text files with the results from a differential expression analysis. It provides
options to output based on FDR and logFC thresholds, positive and negative logFC, and ranked lists.
}
\details{
This function writes out lists of genes to text files. By default, it outputs a list ranked by p-value, lists of genes significant based on FDR and logFC thresholds (all, up, and down).
}
